backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
httpx==0.25.2
python-multipart==0.0.6
prometheus-client==0.19.0
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
coverage==7.3.2